refactor(auth-components): migrate consumers to lib/auth/components and delete old files
LoginPage, signup/page, and join/page now import SignInButton and
EmailPasswordForm from @/lib/auth/components. The old standalone files
(app/auth/login/EmailPasswordForm.tsx, SignInButton.tsx, and their test)
are deleted.
